Quick fix for access-after-free bug
std::stringstream::str().c_str() points to the internal buffer of the
temporary std::string object returned from std::stringstream::str(),
which is deallocated immediately.  So this combination of calls is
illegal.

This commit contains a very stupid but least disruptive fix, which adds
a field to save the std::string object within the Error object.  A
longer-term solution should have what() take a std::ostream & argument
to stuff the string into.